Refactor code into generic NavDecorationWithPrevious
